Fix warning.
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@16899 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sim/sc_sim.cpp
@@ -1676,7 +1676,7 @@ void sim_t::analyze()
   for ( size_t i = 0; i < num_timelines; i++ )
   {
     int last = ( int ) floor( simulation_length.data()[ i ] );
-    assert( last < divisor_timeline.size() ); // We created it with max length
+    assert( last < (int) divisor_timeline.size() ); // We created it with max length
     for ( int j = 0; j <= last; j++ ) divisor_timeline[ j ] += 1;
   }
